requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.28
faiss-cpu==1.8.0
python-dotenv==1.0.1
openai==1.14.2
google-search-results==2.4.2
PyPDF2==3.0.1
tiktoken==0.6.0